+++ pyqode/core/ui/pyqode_core_rc.py
@@ -2,7 +2,7 @@
 
 # Resource object code
 #
-# Created: Sat Aug 31 18:46:36 2013
+# Created: Sun Feb 16 15:48:26 2014
 #      by: The Resource Compiler for PySide (Qt v4.8.4)
 #
 # WARNING! All changes made in this file will be lost!